Renamed files to bookshop for consistency
